requirements.txt
django
django-crispy-forms
crispy-bootstrap5
factory_boy
pillow
sorl-thumbnail